validasi delete image, configredis to cloud

+++ src/modules/movie/movieController.js
@@ -189,8 +189,9 @@ module.exports = {
           null
         );
       }
-
-      deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      if (checkId[0].image) {
+        deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      }
 
       const result = await movieModel.deleteMovie(id);
       return helperWrapper.response(res, 200, "Success delete data", result);
